# packages/contracts/openapi.bundle.yaml
openapi: 3.1.0
info:
  title: Spott API
  version: 1.0.0
  description: Spott iOS/Web 单一契约。具备 Idempotency-Key 参数的副作用写入由调用方提供稳定幂等键； 媒体上传、完成、附件与举报均以调用方保留的 UUID 和服务端持久回执支持响应丢失恢复； 业务状态始终由服务端裁决。
  contact:
    name: Spott Engineering
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: https://api.spott.jp/v1
    description: Production
security:
  - BearerAuth: []
...
paths:
...
  /reports:
    post:
      operationId: createReport
      summary: Create report
      tags:
        - Safety
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportInput'
      responses:
        '201':
          description: Opaque report reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportReceipt'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
  /appeals:
    post:
      operationId: createAppeal
      summary: Create appeal
      tags:
        - Safety
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - statement
              oneOf:
                - required:
                    - caseReference
                - required:
                    - caseId
              properties:
                caseReference:
                  type: string
                  pattern: ^SPT-[0-9]{4}-[A-F0-9]{12}$
                caseId:
                  $ref: '#/components/schemas/UUID'
                statement:
                  type: string
                  minLength: 10
                  maxLength: 5000
      responses:
        '201':
          description: Appeal created or existing appeal returned
        '404':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
  /me/safety-cases:
    get:
      operationId: getMySafetyCases
      summary: List safety reports involving the current user without encrypted evidence
      tags:
        - Safety
      responses:
        '200':
          description: Submitted and subject safety cases
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SafetyCasePage'
        '401':
          $ref: '#/components/responses/Error'
  /me/blocks:
    get:
      operationId: getMyBlocks
      summary: List users blocked by the current user
      tags:
        - Safety
      responses:
        '200':
          description: Blocked users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockPage'
        '401':
          $ref: '#/components/responses/Error'
  /users/{id}/block:
    put:
      operationId: blockUser
      summary: Reach the blocked state and sever mutual follows
      tags:
        - Safety
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  maxLength: 200
      responses:
        '200':
          description: Block state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockState'
        '404':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
    delete:
      operationId: unblockUser
      summary: Reach the unblocked state for a user
      tags:
        - Safety
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Block state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockState'
        '404':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    OpsSession:
      type: apiKey
      in: cookie
      name: __Host-spott_ops
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        oneOf:
          - $ref: '#/components/schemas/UUID'
          - type: string
            maxLength: 100
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        $ref: '#/components/schemas/UUID'
...
  responses:
    Error:
      description: Stable business error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    UUID:
      type: string
      format: uuid
...
    SafetyCasePage:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            type: object
            required:
              - reference
              - relationship
              - targetType
              - targetId
              - reason
              - severity
              - status
              - createdAt
              - updatedAt
            properties:
              reference:
                type: string
                pattern: ^SPT-[0-9]{4}-[A-F0-9]{12}$
              relationship:
                type: string
                enum:
                  - submitted
                  - subject
              targetType:
                type: string
                enum:
                  - event
                  - group
                  - user
                  - comment
                  - announcement
              targetId:
                $ref: '#/components/schemas/UUID'
              reason:
                type: string
              severity:
                type: string
                enum:
                  - p0
                  - p1
                  - p2
              status:
                type: string
              caseStatus:
                type:
                  - string
                  - 'null'
              decision:
                type:
                  - string
                  - 'null'
              slaDueAt:
                type:
                  - string
                  - 'null'
                format: date-time
              createdAt:
                type: string
                format: date-time
              updatedAt:
                type: string
                format: date-time
              appeal:
                oneOf:
                  - type: 'null'
                  - type: object
                    required:
                      - id
                      - status
                      - createdAt
                      - decidedAt
                    properties:
                      id:
                        $ref: '#/components/schemas/UUID'
                      status:
                        type: string
                      createdAt:
                        type:
                          - string
                          - 'null'
                        format: date-time
                      decidedAt:
                        type:
                          - string
                          - 'null'
                        format: date-time
    BlockPage:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            type: object
            required:
              - userId
              - publicHandle
              - blockedAt
            properties:
              userId:
                $ref: '#/components/schemas/UUID'
              publicHandle:
                type: string
              nickname:
                type:
                  - string
                  - 'null'
              reason:
                type:
                  - string
                  - 'null'
              blockedAt:
                type: string
                format: date-time
    BlockState:
      type: object
      required:
        - userId
        - blocked
      properties:
        userId:
          $ref: '#/components/schemas/UUID'
        blocked:
          type: boolean
    ReportReceipt:
      type: object
      additionalProperties: false
      required:
        - reference
        - status
        - submittedAt
      properties:
        reference:
          type: string
          pattern: ^SPT-[0-9]{4}-[A-F0-9]{12}$
        status:
          type: string
          enum:
            - open
        submittedAt:
          type: string
          format: date-time
    ReportInput:
      type: object
      required:
        - targetType
        - targetId
        - reason
      properties:
        targetType:
          type: string
          enum:
            - event
            - group
            - user
            - comment
            - announcement
        targetId:
          $ref: '#/components/schemas/UUID'
        reason:
          type: string
          enum:
            - danger
            - personal_safety
            - fraud
            - harassment
            - harassment_or_hate
            - spam
            - minor_safety
            - other
            - unsafe
        details:
          type: string
          maxLength: 5000
        evidenceAssetIds:
          type: array
          maxItems: 10
          items:
            $ref: '#/components/schemas/UUID'
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - requestId
            - retryable
            - fieldErrors
            - actions
            - meta
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
            retryable:
              type: boolean
            fieldErrors:
              type: array
              items:
                type: object
                required:
                  - field
                  - message
                properties:
                  field:
                    type: string
                  message:
                    type: string
            actions:
              type: array
              items:
                type: object
                required:
                  - type
                  - label
                properties:
                  type:
                    type: string
                  label:
                    type: string
            meta:
              type: object
              additionalProperties: true